Patch by Daniel Furrer -thread implementations  (note; no idea if this actuallay works,.. will see as soon as I can compile and test the whole thing) -fixed a ambiguous function-call
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@12822 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/af/xap/beos/xap_BeOSDlg_FontChooser.cpp
@@ -445,7 +445,7 @@ bool FontWin::QuitRequested()
 
 	// We need to activate the first tab so we can find the view children.
 	BTabView* pTabView = (BTabView *)FindView("fonttab");
-	pTabView->Select(0);
+	pTabView->Select((long int)0);
 	
 	BListView *FontList=(BListView *)FindView("fontlist");
